feat: refactor GitHub calendar and stats components for improved structure and organization
