+++ app/page.tsx
@@ -2,7 +2,7 @@
 
 import Link from "next/link";
 import { useEffect } from "react";
-import GitHubCalendarClient from "@/components/GitHubCalendarClient";
+import GitHubCalendarClient from "@/components/github-calendar-client";
 import { socialLinks, highlights, projects } from "@/lib/constants";
 import AOS from "aos";
 import { event } from "@/lib/gtag";
