+++ components/stats/wakatime-disclaimer.tsx
